Major refactoring Introduced separate view models for each fragment
