+++ app/src/main/java/it/units/borghisegreti/utils/Database.java
@@ -1,0 +1,12 @@
+package it.units.borghisegreti.utils;
+
+public class Database {
+    public static final String DB_URL = "https://adventuremaps-1205-default-rtdb.europe-west1.firebasedatabase.app";
+    public static final String DB_TAG = "FIREBASE_DB";
+    public static final String ZONES_REFERENCE = "zones";
+    public static final String EXPERIENCES_REFERENCE = "experiences";
+    public static final String USER_DATA_REFERENCE = "user_data";
+    public static final String OBJECTIVE_REFERENCE = "objective";
+    public static final String COMPLETED_EXPERIENCES_REFERENCE = "completed_experiences";
+    public static final String POINTS_REFERENCE = "points";
+}
