+++ app/src/main/java/it/units/borghisegreti/viewmodels/MapViewModel.java
@@ -1,5 +1,13 @@
 package it.units.borghisegreti.viewmodels;
 
+import static it.units.borghisegreti.utils.Database.DB_TAG;
+import static it.units.borghisegreti.utils.Database.EXPERIENCES_REFERENCE;
+import static it.units.borghisegreti.utils.Database.ZONES_REFERENCE;
+import static it.units.borghisegreti.utils.Database.COMPLETED_EXPERIENCES_REFERENCE;
+import static it.units.borghisegreti.utils.Database.OBJECTIVE_REFERENCE;
+import static it.units.borghisegreti.utils.Database.POINTS_REFERENCE;
+import static it.units.borghisegreti.utils.Database.USER_DATA_REFERENCE;
+
 import android.util.Log;
 
 import androidx.annotation.NonNull;
@@ -13,6 +21,7 @@ import androidx.lifecycle.ViewModel;
 
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -20,7 +29,10 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 import it.units.borghisegreti.models.Experience;
@@ -28,10 +40,6 @@ import it.units.borghisegreti.models.Zone;
 
 
 public class MapViewModel extends ViewModel {
-    public static final String DB_URL = "https://adventuremaps-1205-default-rtdb.europe-west1.firebasedatabase.app";
-    public static final String DB_TAG = "FIREBASE_DB";
-    public static final String ZONES_REFERENCE = "zones";
-    public static final String EXPERIENCES_REFERENCE = "experiences";
     private static final String CAMERA_LATITUDE = "latitude";
     private static final String CAMERA_LONGITUDE = "longitude";
     private static final String CAMERA_ZOOM = "zoom";
@@ -42,17 +50,30 @@ public class MapViewModel extends ViewModel {
     @NonNull
     private final ValueEventListener zonesListener;
     @NonNull
+    private final ValueEventListener userPointsListener;
+    @NonNull
+    private final ValueEventListener objectiveExperienceListener;
+    @NonNull
     private final MutableLiveData<List<Zone>> databaseZones;
     @NonNull
     private final MutableLiveData<List<Experience>> databaseExperiences;
     @NonNull
+    private final MutableLiveData<Integer> databaseUserPoints;
+    @NonNull
+    private final MutableLiveData<String> databaseObjectiveExperienceId;
+    @NonNull
     private final SavedStateHandle savedState;
+    @NonNull
+    private final String userId;
 
     public MapViewModel(@NonNull FirebaseDatabase database, @NonNull SavedStateHandle savedState) {
         this.savedState = savedState;
         this.database = database;
         databaseZones = new MutableLiveData<>();
         databaseExperiences = new MutableLiveData<>();
+        databaseUserPoints = new MutableLiveData<>();
+        databaseObjectiveExperienceId = new MutableLiveData<>();
+        userId = Objects.requireNonNull(FirebaseAuth.getInstance().getUid(), "User should be already authenticated");
         experiencesListener = new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
@@ -87,8 +108,40 @@ public class MapViewModel extends ViewModel {
                 Log.e(DB_TAG, "Error while getting new zones: " + error.getMessage());
             }
         };
+        userPointsListener = new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                Integer points = snapshot.getValue(Integer.class);
+                databaseUserPoints.setValue(points);
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+                Log.e(DB_TAG, "Error: " + error.getMessage(), error.toException());
+            }
+        };
+        objectiveExperienceListener = new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                String objectiveExperienceId = snapshot.getValue(String.class);
+                databaseObjectiveExperienceId.setValue(objectiveExperienceId);
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+                Log.e(DB_TAG, "Error: " + error.getMessage(), error.toException());
+            }
+        };
         database.getReference(ZONES_REFERENCE).addValueEventListener(zonesListener);
         database.getReference(EXPERIENCES_REFERENCE).addValueEventListener(experiencesListener);
+        database.getReference(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(POINTS_REFERENCE)
+                .addValueEventListener(userPointsListener);
+        database.getReference(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(OBJECTIVE_REFERENCE)
+                .addValueEventListener(objectiveExperienceListener);
     }
 
     /**
@@ -152,6 +205,51 @@ public class MapViewModel extends ViewModel {
         return experienceReference.setValue(experience);
     }
 
+    @NonNull
+    public Task<Void> setExperienceAsCompleted(@NonNull Experience experience) {
+        experience.setDateOfCompletion(new Date());
+        setUserPoints(experience.getPoints() + Objects.requireNonNull(databaseUserPoints.getValue(), "User points should already be initialized")).addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                Log.d(DB_TAG, "Successfully updated user points");
+            } else {
+                Log.e(DB_TAG, "Error while updating user points");
+            }
+        });
+        Map<String, Object> completedExperienceById = new HashMap<>();
+        completedExperienceById.put(experience.getId(), experience);
+        return database.getReference()
+                .child(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(COMPLETED_EXPERIENCES_REFERENCE)
+                .updateChildren(completedExperienceById);
+    }
+
+    @NonNull
+    private Task<Void> setUserPoints(int points) {
+        return database.getReference()
+                .child(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(POINTS_REFERENCE)
+                .setValue(points);
+    }
+
+    @NonNull
+    public Task<Void> setObjectiveExperience(@Nullable String experienceId) {
+        return database.getReference()
+                .child(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(OBJECTIVE_REFERENCE)
+                .setValue(experienceId);
+    }
+
+    /**
+     *
+     * @return The objective experience's ID for the current user, or null if the user has no objective experience set
+     */
+    public LiveData<String> getObjectiveExperienceId() {
+        return databaseObjectiveExperienceId;
+    }
+
     @Nullable
     public Double getCameraLatitude() {
         return savedState.get(CAMERA_LATITUDE);
@@ -184,6 +282,14 @@ public class MapViewModel extends ViewModel {
         super.onCleared();
         database.getReference(ZONES_REFERENCE).removeEventListener(zonesListener);
         database.getReference(EXPERIENCES_REFERENCE).removeEventListener(experiencesListener);
+        database.getReference(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(POINTS_REFERENCE)
+                .removeEventListener(userPointsListener);
+        database.getReference(USER_DATA_REFERENCE)
+                .child(userId)
+                .child(OBJECTIVE_REFERENCE)
+                .removeEventListener(objectiveExperienceListener);
     }
 
     public static class Factory extends AbstractSavedStateViewModelFactory {
